Edit schemas

+++ server/db/postgresListing.sql
@@ -3,4 +3,4 @@ CREATE DATABASE listing;
 \c listing;
 
 CREATE TABLE booking (listing_id text, basePrice smallint, views smallint, cleaningFee smallint, serviceFee smallint, taxes smallint,
-maxGuests smallint, lastAvailableDate date);
+maxGuests tinyint, lastAvailableDate date);
